requirement.txt
aniso8601==9.0.1
APScheduler==3.10.4
async-timeout==4.0.3
backports.zoneinfo==0.2.1
bcrypt==4.1.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.3.0
Flask==2.2.5
Flask-APScheduler==1.12.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-language==2.13.3
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
newsapi-python==0.2.7
pkg_resources==0.0.0
proto-plus==1.23.0
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.7.1
tzlocal==5.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
